+++ docs/CHANGE_LOG.md
@@ -5,6 +5,9 @@
 #### **Binding Options - Custom Triggers:**
 - Added a new binding option called "onRefresh", which states an event that should be triggered when a heatmap is refreshed.
 
+#### **Public Functions:**
+- Added new public function "removeDate()", which will remove a specific date from the heat map (or decrease the date if more than one exists).
+
 <br>